+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 **NOTE:** The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+##### [Unreleased]
+
+**CHANGED**
+
+* Exit when connection to MQTT broker fails.
+
+
 ##### [1.0.2] - 2019-11-11
 
 **SECURITY**